TS/requirements.txt
Django~=3.2.5
djangorestframework>=3.11.0,<3.13
uWSGI>=2.0.19.1,<2.1
gunicorn>=20.1.0,<20.3
Pillow>=8.1.2,<8.4
psycopg2>=2.8.6,<2.9
django-cors-headers>=3.7.0,<3.9
python-dotenv>=0.16.0,<0.19
channels>=3.0.3,<3.2
channels-redis>=3.2.0,<3.5
###dj-database-url>=0.5.0,0.9
whitenoise>=5.2.0,<5.8
django-storages>=1.11.1,<1.19
boto3>=1.17.40,<1.19
django-filter>=2.4.0,<2.8
dj-rest-auth[with_social]
###dj-rest-auth>=2.1.2,<2.5
###django-allauth>=0.44.0,<0.6
drf-yasg2>=1.19.4,<1.19.9
rest-framework-generic-relations>=2.0.0,<2.4
djangorestframework-simplejwt>=4.7.1,<4.7.9
django-cleanup>=5.1.0,<5.7
###geopy>=2.1.0,<2.5
geoip2>=4.1.0,<4.9
python-decouple>=3.4.0,<3.49
##python-libmagic==0.4.0
django-upload-validator>=1.1.5,<1.1.9
django-import-export>=2.6.1,<2.6.8
django-ckeditor>=6.1.0,<6.1.7
PyJWT==2.1.0

